New zifferjs documentation
