+++ src/InterfaceLogic.ts
@@ -502,7 +502,11 @@ export const installInterfaceLogic = (app: Editor) => {
     "synths",
     "chaining",
     "patterns",
-    "ziffers",
+    "ziffers_basics",
+    "ziffers_scales",
+    "ziffers_rhythm",
+    "ziffers_algorithmic",
+    "ziffers_tonnetz",
     "midi",
     "osc",
     "functions",
